+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/SysOutboundGoods.java
@@ -7,102 +7,106 @@ import com.ruoyi.common.core.domain.BaseEntity;
 
 /**
  * 【请填写功能名称】对象 sys_outbound_goods
- * 
+ *
  * @author ruoyi
  * @date 2024-03-17
  */
-public class SysOutboundGoods extends BaseEntity
-{
+public class SysOutboundGoods extends BaseEntity {
     private static final long serialVersionUID = 1L;
 
-    /** 出库记录id */
+    /**
+     * 出库记录id
+     */
     @Excel(name = "出库记录id")
     private Long outboundId;
 
-    /** 商品id */
+    /**
+     * 商品id
+     */
     @Excel(name = "商品id")
     private Long goodsId;
 
-    /** 数量 */
+    /**
+     * 数量
+     */
     @Excel(name = "数量")
     private Long count;
 
-    /** 总价格 */
+    /**
+     * 总价格
+     */
     @Excel(name = "总价格")
-    private Long priceTotal;
+    private Double priceTotal;
 
-    /** 单价 */
+    /**
+     * 单价
+     */
     @Excel(name = "单价")
-    private Long price;
+    private Double price;
 
-    /** 用框数 */
+    /**
+     * 用框数
+     */
     @Excel(name = "用框数")
     private Long caseTotal;
 
-    public void setOutboundId(Long outboundId) 
-    {
+    public void setOutboundId(Long outboundId) {
         this.outboundId = outboundId;
     }
 
-    public Long getOutboundId() 
-    {
+    public Long getOutboundId() {
         return outboundId;
     }
-    public void setGoodsId(Long goodsId) 
-    {
+
+    public void setGoodsId(Long goodsId) {
         this.goodsId = goodsId;
     }
 
-    public Long getGoodsId() 
-    {
+    public Long getGoodsId() {
         return goodsId;
     }
-    public void setCount(Long count) 
-    {
+
+    public void setCount(Long count) {
         this.count = count;
     }
 
-    public Long getCount() 
-    {
+    public Long getCount() {
         return count;
     }
-    public void setPriceTotal(Long priceTotal) 
-    {
+
+    public void setPriceTotal(Double priceTotal) {
         this.priceTotal = priceTotal;
     }
 
-    public Long getPriceTotal() 
-    {
+    public Double getPriceTotal() {
         return priceTotal;
     }
-    public void setPrice(Long price) 
-    {
+
+    public void setPrice(Double price) {
         this.price = price;
     }
 
-    public Long getPrice() 
-    {
+    public Double getPrice() {
         return price;
     }
-    public void setCaseTotal(Long caseTotal) 
-    {
+
+    public void setCaseTotal(Long caseTotal) {
         this.caseTotal = caseTotal;
     }
 
-    public Long getCaseTotal() 
-    {
+    public Long getCaseTotal() {
         return caseTotal;
     }
 
     @Override
     public String toString() {
-        return new ToStringBuilder(this,ToStringStyle.MULTI_LINE_STYLE)
-            .append("outboundId", getOutboundId())
-            .append("goodsId", getGoodsId())
-            .append("count", getCount())
-            .append("priceTotal", getPriceTotal())
-            .append("price", getPrice())
-            .append("caseTotal", getCaseTotal())
-            .toString();
+        return new ToStringBuilder(this, ToStringStyle.MULTI_LINE_STYLE)
+                .append("outboundId", getOutboundId())
+                .append("goodsId", getGoodsId())
+                .append("count", getCount())
+                .append("priceTotal", getPriceTotal())
+                .append("price", getPrice())
+                .append("caseTotal", getCaseTotal())
+                .toString();
     }
 }
